+++ examples/src/app/ApplicationAnt.tsx
@@ -16,6 +16,7 @@ import {
     OneOfAnt,
     OneOfButtonFormAnt,
     OneOfSelectAnt,
+    ProgressFormAnt,
     RateFormAnt,
     TableAnt,
     TagAnt,
@@ -155,6 +156,17 @@ export class ApplicationAnt extends React.Component<{ app?: Application }> {
                             <br />
                             <h3>SliderAnt Component</h3>
                             <SliderFormAnt operation={application.testSliderNumeric} showNumber={true} />
+                            <br />
+                            <h3>ProgressAnt Component</h3>
+                            <ProgressFormAnt operation={application.testProgress} />
+                            <ActionButtonAnt operation={application.decreaseLineProgress} />
+                            <ActionButtonAnt operation={application.increaseLineProgress} />
+                            <br />
+                            <br />
+                            <ProgressFormAnt operation={application.testProgress2} />
+                            <ActionButtonAnt operation={application.decreaseCircleProgress} />
+                            <ActionButtonAnt operation={application.increaseCircleProgress} />
+                            <br />
                             <br />
                             <h3>TagAnt Component</h3>
                             <TagAnt operation={application.testTags} tagColor="#87d068" newItemLabel={'tstst'} />